Allow DataPoint to be created from a TimeSeriesName object (#41)
In certain scenarios where we need to send a set of time series with
fixed metric names and tags, the old send API requires creating
DataPoint object for each point in those series. This can quickly turn
out to be very expensive and slow down the code. We introduce a new
send_timeseries API which accepts a list of TimeSeries objects.

This requires adding an add_point function to the TimeSeries class and
refactoring some of the common code for creating the payload from dps
list or series list.

We also move the encode metric/decode metric functions to TimeSeries
class and deprecate the timeseries module