Reporter send interval can become 0 if the process is paused for a while(#56)
In case the process is paused, and wakes up a while later (may be hours), the way the next reporting time was being calculated, would mean that the reporter would keep sending the data much more frequently than the reporting interval set. We should calculate the time it took to send the data and then calculate the difference from the reporting interval to decide how much to sleep.